requirements.txt
anyio==4.4.0
autopep8==2.3.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
fastcore==1.6.1
fastlite==0.0.7
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
pycodestyle==2.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.2.0
python-multipart==0.0.9
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlite_minutils==3.36.0.post4
starlette==0.38.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
